salvataggio

+++ c04/ex03/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: fgolino <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 01:07:01 by fgolino           #+#    #+#             */
-/*   Updated: 2022/10/26 09:09:26 by fgolino          ###   ########.fr       */
+/*   Updated: 2022/10/30 11:10:38 by fgolino          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	loop(char *str, int i)
 {
-	int	nb;
+	long	nb;
 
 	nb = 0;
 	while (str[i] < 58 && str[i] > 47 && str[i] != 0)
@@ -20,14 +20,14 @@ int	loop(char *str, int i)
 		nb = (nb * 10) + (str[i] - 48);
 		i++;
 	}
-	return (nb);
+	return ((int)nb);
 }
 
 int	ft_atoi(char *str)
 {
 	int	i;
 	int	sign;
-	int	nb;
+	long	nb;
 
 	i = 0;
 	sign = 1;
@@ -43,7 +43,7 @@ int	ft_atoi(char *str)
 		else if (str[i] < 58 && str[i] > 47)
 		{
 			nb = loop(str, i);
-			return (nb * sign);
+			return (int(nb * sign));
 		}
 		else
 			i++;
